+++ APSI/apsi/plaintextarith.cpp
@@ -9,7 +9,7 @@ using namespace seal::util;
 
 namespace apsi
 {
-    ConstPointer<std::uint64_t> duplicate_poly_if_needed(const Plaintext &poly, int new_coeff_count, bool force, MemoryPool &pool)
+    ConstPointer<std::uint64_t> duplicate_poly_if_needed(const Plaintext &poly, size_t new_coeff_count, bool force, MemoryPool &pool)
     {
         return util::duplicate_poly_if_needed(poly.data(), poly.coeff_count(), 1, new_coeff_count, 1, force, pool);
     }
@@ -19,9 +19,9 @@ namespace apsi
         return util::duplicate_if_needed(poly.data(), poly.coeff_count() * poly.coeff_uint64_count(), force, pool);
     }
 
-    void resize_destination_if_needed(Plaintext &destination, int coeff_count)
+    void resize_destination_if_needed(Plaintext &destination, size_t coeff_count)
     {
-        int dest_coeff_count = destination.coeff_count();
+        size_t dest_coeff_count = destination.coeff_count();
         if (dest_coeff_count < coeff_count)
         {
             destination.resize(coeff_count);
@@ -35,9 +35,9 @@ namespace apsi
     void multiply(const seal::Plaintext &plaintext1, const seal::Plaintext &plaintext2,
         const seal::util::PolyModulus &poly_mod, const seal::SmallModulus &coeff_mod, seal::Plaintext &result, MemoryPoolHandle pool)
     {
-        int coeff_count = poly_mod.coeff_count();
+        size_t coeff_count = poly_mod.coeff_count();
         int coeff_bit_count = coeff_mod.bit_count();
-        int coeff_uint64_count = coeff_mod.uint64_count();
+        size_t coeff_uint64_count = coeff_mod.uint64_count();
 
         // Get pointer to inputs (duplicated and resized if needed).
         auto poly1ptr = duplicate_poly_if_needed(plaintext1, coeff_count, plaintext1.data() == result.data(), pool);
@@ -54,9 +54,9 @@ namespace apsi
         const seal::util::PolyModulus &poly_mod, const seal::SmallModulus &coeff_mod, seal::Plaintext &result, seal::MemoryPoolHandle pool)
     {
         // Verify parameters.
-        int coeff_count = poly_mod.coeff_count();
+        size_t coeff_count = poly_mod.coeff_count();
         int coeff_bit_count = coeff_mod.bit_count();
-        int coeff_uint64_count = coeff_mod.uint64_count();
+        size_t coeff_uint64_count = coeff_mod.uint64_count();
 
         // Get pointer to inputs (duplicated and resized if needed).
         auto poly1ptr = duplicate_poly_if_needed(plaintext1, coeff_count, plaintext1.data() == result.data(), pool);
@@ -73,9 +73,9 @@ namespace apsi
         const seal::util::PolyModulus &poly_mod, const seal::SmallModulus &coeff_mod, seal::Plaintext &result, seal::MemoryPoolHandle pool)
     {
         // Verify parameters.
-        int coeff_count = poly_mod.coeff_count();
+        size_t coeff_count = poly_mod.coeff_count();
         int coeff_bit_count = coeff_mod.bit_count();
-        int coeff_uint64_count = coeff_mod.uint64_count();
+        size_t coeff_uint64_count = coeff_mod.uint64_count();
 
         // Get pointer to inputs (duplicated and resized if needed).
         auto poly1ptr = duplicate_poly_if_needed(plaintext1, coeff_count, plaintext1.data() == result.data(), pool);
